+++ server/controllers/convenios.controles.js
@@ -3,8 +3,12 @@ import { pool } from "../db.js";
 
 async function getConvenios(req, res) {
   try {
-    const [result] = await pool.query("SELECT name, seccion, ruta_logo, descripcion FROM convenio");
-    res.json(result);
+    if (req.session.authenticated) {
+      const [result] = await pool.query(
+        "SELECT name, seccion, ruta_logo, descripcion FROM convenio"
+      );
+      res.json(result);
+    }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
   }
@@ -12,14 +16,16 @@ async function getConvenios(req, res) {
 
 async function getoneConvenio(req, res) {
   try {
-    const [result] = await pool.query(
-      "SELECT * FROM convenio WHERE convenio_id = ?",
-      req.params.id
-    );
-    if (result.length === 0) {
-      return res.status(403).json({ Message: "Convenio not found" });
-    } else {
-      res.json(result[0]);
+    if (req.session.authenticated) {
+      const [result] = await pool.query(
+        "SELECT * FROM convenio WHERE convenio_id = ?",
+        req.params.id
+      );
+      if (result.length === 0) {
+        return res.status(403).json({ Message: "Convenio not found" });
+      } else {
+        res.json(result[0]);
+      }
     }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
@@ -28,15 +34,17 @@ async function getoneConvenio(req, res) {
 
 async function createConvenio(req, res) {
   try {
-    const { name, seccion, created_by, ruta_logo, descripcion } = req.body;
-    const [result] = await pool.query(
-      "INSERT INTO convenio (name, seccion, created_by, ruta_logo, descripcion) VALUES (?, ?, ?, ?, ?)",
-      [name, seccion, created_by, ruta_logo, descripcion]
-    );
-    res.json({
-      id: result.insertId,
-      message: `Convenio creado exitosamente`,
-    });
+    if (req.session.user.es_admin) {
+      const { name, seccion, ruta_logo, descripcion } = req.body;
+      const [result] = await pool.query(
+        "INSERT INTO convenio (name, seccion, created_by, ruta_logo, descripcion) VALUES (?, ?, ?, ?, ?)",
+        [name, seccion, req.session.user.user_id, ruta_logo, descripcion]
+      );
+      res.json({
+        id: result.insertId,
+        message: `Convenio creado exitosamente`,
+      });
+    }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -44,13 +52,15 @@ async function createConvenio(req, res) {
 
 async function uptadateConvenio(req, res) {
   try {
-    const [result] = await pool.query(
-      "UPDATE convenio SET ? WHERE convenio_id = ?",
-      [req.body, req.params.id]
-    );
-    if (result.affectedRows === 0)
-      return res.status(403).json({ message: "No se actualizo ningún dato" });
-    res.json(result);
+    if (req.session.user.es_admin) {
+      const [result] = await pool.query(
+        "UPDATE convenio SET ? WHERE convenio_id = ?",
+        [req.body, req.params.id]
+      );
+      if (result.affectedRows === 0)
+        return res.status(403).json({ message: "No se actualizo ningún dato" });
+      res.json(result);
+    }
   } catch (error) {
     console.log(error);
     return res.status(500).json({ message: console.error.message });
@@ -59,13 +69,17 @@ async function uptadateConvenio(req, res) {
 
 async function deleteConvenio(req, res) {
   try {
-    const [result] = await pool.query(
-      "DELETE FROM convenio WHERE convenio_id = ?",
-      req.params.id
-    );
-    if (result.affectedRows === 0)
-      return res.status(403).json({ message: "Convenio no encontrado" });
-    return res.status(204).json({ message: "Convenio eliminada exitosamente" });
+    if (req.session.user.es_admin) {
+      const [result] = await pool.query(
+        "DELETE FROM convenio WHERE convenio_id = ?",
+        req.params.id
+      );
+      if (result.affectedRows === 0)
+        return res.status(403).json({ message: "Convenio no encontrado" });
+      return res
+        .status(204)
+        .json({ message: "Convenio eliminada exitosamente" });
+    }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
   }
@@ -75,14 +89,16 @@ async function deleteConvenio(req, res) {
 
 async function getFilterConv(req, res) {
   try {
-    const [result] = await pool.query(
-      "SELECT * FROM convenio WHERE seccion = ?",
-      req.params.id
-    );
-    if (result.length === 0) {
-      return res.status(403).json({ Message: "Secciones no encontradas" });
-    } else {
-      res.json(result);
+    if (req.session.authenticated) {
+      const [result] = await pool.query(
+        "SELECT * FROM convenio WHERE seccion = ?",
+        req.params.id
+      );
+      if (result.length === 0) {
+        return res.status(403).json({ Message: "Secciones no encontradas" });
+      } else {
+        res.json(result);
+      }
     }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
