+++ server/controllers/news.controles.js
@@ -3,10 +3,12 @@ import { pool } from "../db.js";
 
 async function getNews(req, res) {
   try {
-    const [result] = await pool.query(
-      "SELECT noticia_id, name, fecha, ruta_imagen FROM noticias"
-    );
-    res.json(result);
+    if (req.session.authenticated) {
+      const [result] = await pool.query(
+        "SELECT noticia_id, name, fecha, ruta_imagen FROM noticias"
+      );
+      res.json(result);
+    }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
   }
@@ -14,14 +16,16 @@ async function getNews(req, res) {
 
 async function getoneNew(req, res) {
   try {
-    const [result] = await pool.query(
-      "SELECT * FROM noticias WHERE noticia_id = ?",
-      req.params.id
-    );
-    if (result.length === 0) {
-      return res.status(403).json({ Message: "Noticia not found" });
-    } else {
-      res.json(result[0]);
+    if (req.session.authenticated) {
+      const [result] = await pool.query(
+        "SELECT * FROM noticias WHERE noticia_id = ?",
+        req.params.id
+      );
+      if (result.length === 0) {
+        return res.status(403).json({ Message: "Noticia not found" });
+      } else {
+        res.json(result[0]);
+      }
     }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
@@ -30,15 +34,17 @@ async function getoneNew(req, res) {
 
 async function createNew(req, res) {
   try {
-    const { name, ruta_imagen, descripcion, created_by } = req.body;
-    const [result] = await pool.query(
-      "INSERT INTO noticias (name, fecha, ruta_imagen, descripcion, created_by) VALUES (?, NOW(), ?, ?, ?)",
-      [name, ruta_imagen, descripcion, created_by]
-    );
-    res.json({
-      id: result.insertId,
-      message: `Noticia creada exitosamente`,
-    });
+    if (req.session.user.es_admin) {
+      const { name, ruta_imagen, descripcion } = req.body;
+      const [result] = await pool.query(
+        "INSERT INTO noticias (name, fecha, ruta_imagen, descripcion, created_by) VALUES (?, NOW(), ?, ?, ?)",
+        [name, ruta_imagen, descripcion, req.session.user.user_id]
+      );
+      res.json({
+        id: result.insertId,
+        message: `Noticia creada exitosamente`,
+      });
+    }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -46,13 +52,15 @@ async function createNew(req, res) {
 
 async function uptadateNews(req, res) {
   try {
-    const [result] = await pool.query(
-      "UPDATE noticias SET ? WHERE noticia_id = ?",
-      [req.body, req.params.id]
-    );
-    if (result.affectedRows === 0)
-      return res.status(403).json({ message: "No se actualizo ningún dato" });
-    res.json(result);
+    if (req.session.user.es_admin) {
+      const [result] = await pool.query(
+        "UPDATE noticias SET ? WHERE noticia_id = ?",
+        [req.body, req.params.id]
+      );
+      if (result.affectedRows === 0)
+        return res.status(403).json({ message: "No se actualizo ningún dato" });
+      res.json(result);
+    }
   } catch (error) {
     console.log(error);
     return res.status(500).json({ message: console.error.message });
@@ -61,13 +69,15 @@ async function uptadateNews(req, res) {
 
 async function deleteNew(req, res) {
   try {
-    const [result] = await pool.query(
-      "DELETE FROM noticias WHERE noticia_id = ?",
-      req.params.id
-    );
-    if (result.affectedRows === 0)
-      return res.status(403).json({ message: "Noticia no encontrada" });
-    return res.status(204).json({ message: "Noticia eliminada exitosamente" });
+    if (req.session.user.es_admin) {
+      const [result] = await pool.query(
+        "DELETE FROM noticias WHERE noticia_id = ?",
+        req.params.id
+      );
+      if (result.affectedRows === 0)
+        return res.status(403).json({ message: "Noticia no encontrada" });
+      res.json({ message: "Noticia eliminada exitosamente" });
+    }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
   }
