+++ server/controllers/retocoments.controles.js
@@ -3,15 +3,17 @@ import { pool } from "../db.js";
 
 async function createRetoComent(req, res) {
   try {
-    const { reto_id, user_id, contenido } = req.body;
-    const [result] = await pool.query(
-      "INSERT INTO reto_coments (reto_id, user_id, contenido, fecha) VALUES (?, ?, ?, NOW())",
-      [reto_id, user_id, contenido]
-    );
-    res.json({
-      id: result.insertId,
-      message: `Comentario creado exitosamente`,
-    });
+    if (req.session.authenticated) {
+      const { contenido } = req.body;
+      const [result] = await pool.query(
+        "INSERT INTO reto_coments (reto_id, user_id, contenido, fecha) VALUES (?, ?, ?, NOW())",
+        [req.params.aventura_id, req.session.user.user_id, contenido]
+      );
+      res.json({
+        id: result.insertId,
+        message: `Comentario creado exitosamente`,
+      });
+    }
   } catch (error) {
     return res.status(500).json({ message: error.message });
   }
@@ -19,13 +21,22 @@ async function createRetoComent(req, res) {
 
 async function uptadateRetoComent(req, res) {
   try {
-    const [result] = await pool.query(
-      "UPDATE reto_coments SET ? WHERE coment_id = ?",
-      [req.body, req.params.id]
+    const [post] = await pool.query(
+      "SELECT user_id FROM reto_coments WHERE coment_id = ?",
+      [req.params.id]
     );
-    if (result.affectedRows === 0)
-      return res.status(403).json({ message: "No se actualizo ningún dato" });
-    res.json(result);
+    if (
+      req.session.user.user_id == post[0].user_id ||
+      req.session.user.es_admin
+    ) {
+      const [result] = await pool.query(
+        "UPDATE reto_coments SET ? WHERE coment_id = ?",
+        [req.body, req.params.id]
+      );
+      if (result.affectedRows === 0)
+        return res.status(403).json({ message: "No se actualizo ningún dato" });
+      res.json(result);
+    }
   } catch (error) {
     console.log(error);
     return res.status(500).json({ message: console.error.message });
@@ -34,15 +45,24 @@ async function uptadateRetoComent(req, res) {
 
 async function deleteRetoComent(req, res) {
   try {
-    const [result] = await pool.query(
-      "DELETE FROM reto_coments WHERE coment_id = ?",
-      req.params.id
+    const [post] = await pool.query(
+      "SELECT user_id FROM reto_coments WHERE coment_id = ?",
+      [req.params.id]
     );
-    if (result.affectedRows === 0)
-      return res.status(403).json({ message: "Comentario no encontrado" });
-    return res
-      .status(204)
-      .json({ message: "Comentario eliminado exitosamente" });
+    if (
+      req.session.user.user_id == post[0].user_id ||
+      req.session.user.es_admin
+    ) {
+      const [result] = await pool.query(
+        "DELETE FROM reto_coments WHERE coment_id = ?",
+        req.params.id
+      );
+      if (result.affectedRows === 0)
+        return res.status(403).json({ message: "Comentario no encontrado" });
+      return res
+        .status(204)
+        .json({ message: "Comentario eliminado exitosamente" });
+    }
   } catch (error) {
     return res.status(500).json({ message: console.error.message });
   }
